refactor(compiler-cli): extract `ModuleWithProviders` analysis out of NgModule handler
This commit moves the foreign function resolver logic for detecting a
`ModuleWithProviders` in a return type position of a function call, as the logic can
then be reused for standalone components in a subsequent commit.
